pyproject.toml
[project]
name = "linguistic-stratigraphy"
version = "0.1.0"
description = "Automated infrastructure for building and maintaining a cross-linguistic lexical evolution graph"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Linguistic Stratigraphy Team" }
]
keywords = [
    "linguistics",
    "etymology",
    "nlp",
    "knowledge-graph",
    "diachronic",
    "historical-linguistics",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
    "Typing :: Typed",
]

dependencies = [
    # Core
    "pydantic>=2.0",
    "httpx>=0.24",
    "aiohttp>=3.8",

    # Databases
    "neo4j>=5.0",
    "asyncpg>=0.28",
    "elasticsearch>=8.0",
    "redis>=4.0",
    "pymilvus>=2.3",

    # NLP
    "spacy>=3.6",
    "stanza>=1.5",
    "sentence-transformers>=2.2",
    "transformers>=4.30",

    # ML
    "scikit-learn>=1.3",
    "xgboost>=1.7",
    "torch>=2.0",
    "numpy>=1.24",

    # API
    "fastapi>=0.100",
    "uvicorn>=0.23",
    "strawberry-graphql>=0.200",

    # Orchestration
    "apache-airflow>=2.7",

    # Utils
    "pywikibot>=8.0",
    "lxml>=4.9",
    "tqdm>=4.65",
    "python-Levenshtein>=0.21",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.1",
    "pytest-xdist>=3.3",
    "hypothesis>=6.82",

    # Code quality
    "black>=24.1",
    "ruff>=0.1.14",
    "mypy>=1.8",
    "pre-commit>=3.6",

    # Security
    "bandit>=1.7",

    # Type stubs
    "types-requests>=2.31",
    "types-redis>=4.6",
]

docs = [
    "mkdocs>=1.5",
    "mkdocs-material>=9.4",
    "mkdocstrings[python]>=0.23",
]

[project.urls]
Homepage = "https://github.com/linguistic-stratigraphy/linguistic-stratigraphy"
Documentation = "https://linguistic-stratigraphy.readthedocs.io"
Repository = "https://github.com/linguistic-stratigraphy/linguistic-stratigraphy"
Issues = "https://github.com/linguistic-stratigraphy/linguistic-stratigraphy/issues"

[project.scripts]
ls-api = "src.api.main:run"
ls-ingest = "src.adapters.wiktionary:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

# Black configuration
[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.mypy_cache
    | \.ruff_cache
    | \.venv
    | venv
    | __pycache__
    | build
    | dist
)/
'''

# mypy configuration
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "silent"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false

# pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "-ra",
]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

# Coverage configuration
[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
fail_under = 80
show_missing = true

# Ruff linter configuration
[tool.ruff]
line-length = 100
target-version = "py311"
exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
    "venv",
    "__pycache__",
    "build",
    "dist",
]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "SIM",  # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function call in default argument (needed for FastAPI Depends)
    "SIM102", # collapsible-if (sometimes less readable)
    "UP007",  # use X | Y for Union (pydantic compatibility)
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

# Bandit security configuration
[tool.bandit]
exclude_dirs = ["tests", "venv", ".venv"]
skips = [
    "B101",  # Allow assert in non-test code for now
    "B104",  # Binding to 0.0.0.0 is intentional for containerized deployments
]
